pyproject.toml
[project]
name = "elephant-gun"
version = "0.1.0"
description = "Elephant Gun 🐘🔫 - Hybrid SQL + semantic search CLI for Postgres (local-first)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "sentence-transformers>=3.0.0",
  "psycopg[binary]>=3.1.18",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
elephant-gun = "elephant_gun:main"
egun = "elephant_gun:main"
